api_locale/requirements.txt
Django==2.2
pyOpenSSL==17.5.0
pyparsing==2.1.4
pytest==3.0.5
python-dateutil==2.6.0
gunicorn
django-heroku
psycopg2-binary
dj-database-url
djangorestframework
